made changes

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -12,8 +12,8 @@
 
 size_t print_list(const list_t *h)
 {
-	int len;
-	int elem_count;
+	size_t len;
+	size_t elem_count;
 	char *string;
 
 	elem_count = 0;
@@ -35,7 +35,7 @@ size_t print_list(const list_t *h)
 			string = "(nil)";
 		}
 
-		printf("[%d] %s\n", len, string);
+		printf("[%ld] %s\n", len, string);
 
 		elem_count += 1;
 		h = h->next;
